+++ internal/ascii/ascii.go
